fix genType tests
