+++ tests/gentype_tests/typescript-react-example/src/MyInput.gen.tsx
@@ -11,7 +11,7 @@ export const defaultTypeChecked: React.ComponentType<{ readonly onFocus?: (_1:in
 // Export '$$default' early to allow circular import from the '.bs.js' file.
 export const $$default: unknown = defaultTypeChecked as React.ComponentType<{ readonly onFocus?: (_1:inputFocusEvent) => void }> as any;
 
-import type {inputFocusEvent as $$inputFocusEvent} from './shims/ReactEvent.shim';
+import type {inputFocusEvent as $$inputFocusEvent} from './shims/JsxEvent.shim';
 
 export type inputFocusEvent = $$inputFocusEvent;
