+++ tests/gentype_tests/typescript-react-example/src/shims/JsxEvent.shim.ts
@@ -1,3 +1,61 @@
+export abstract class Animation_t {
+  protected opaque: unknown;
+}
+
+export abstract class Clipboard_t {
+  protected opaque: unknown;
+}
+
+export abstract class Composition_t {
+  protected opaque: unknown;
+}
+
+export abstract class Focus_t {
+  protected opaque: unknown;
+}
+
+export abstract class Form_t {
+  protected opaque: unknown;
+}
+
+export abstract class Keyboard_t {
+  protected opaque: unknown;
+}
+
+export abstract class Image_t {
+  protected opaque: unknown;
+}
+
+export abstract class Media_t {
+  protected opaque: unknown;
+}
+
+export abstract class Selection_t {
+  protected opaque: unknown;
+}
+
+export abstract class Synthetic_t {
+  protected opaque: unknown;
+}
+
+export abstract class Touch_t {
+  protected opaque: unknown;
+}
+
+export abstract class Transition_t {
+  protected opaque: unknown;
+}
+
+export abstract class UI_t {
+  protected opaque: unknown;
+}
+
+export abstract class Wheel_t {
+  protected opaque: unknown;
+}
+
+export type inputFocusEvent = React.FocusEvent<HTMLInputElement>;
+
 export abstract class Mouse_t {
   protected opaque: unknown;
 }
